+++ cmd/backtest_tool/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "fmt"
+
+func main() {
+	fmt.Println("Backtest tool - Placeholder")
+}
