+++ internal/strategy/scalping.go
@@ -1,0 +1,1 @@
+package strategy
