+++ internal/strategy/swing.go
@@ -1,0 +1,1 @@
+package strategy
